COM-1630 move ProfileHeader to core
